requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# LlamaIndex core
llama-index>=0.10.0
llama-index-core>=0.10.0
llama-index-embeddings-openai>=0.1.0
llama-index-embeddings-gemini>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-openai>=0.1.0
llama-index-readers-file>=0.1.0

# LLM providers
openai>=1.10.0
google-generativeai>=0.3.0

# Document processing
pypdf>=3.17.0
docx2txt>=0.8

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
aiohttp>=3.9.0

# Slack (optional, for direct integration)
slack-sdk>=3.26.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
isort>=5.13.0
